content-ranking-system/requirements.txt
numpy
pandas
pyyaml
scikit-learn
lightgbm
fastapi
uvicorn
pydantic
pytest


